refactor: Change package to pieces
- Pawn's constructor that takes String type of color as parameter caused compile error since the constructor was not public.
