+++ src/test/java/pieces/PawnTest.java
@@ -1,7 +1,9 @@
-package chess;
+package pieces;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import pieces.Pawn;
+
 import static org.assertj.core.api.Assertions.*;
